Fix double precision for most routines, keeping kind(dp) for memory intensive arrays.

+++ amr/boundaries.f90
@@ -107,11 +107,11 @@ subroutine init_bound_refine(r,g,m,grid,grid_ref,ibound)
 
   integer::idim, ind, ivar, irad
   integer::type, dir, shift, nstride
-  real(dp)::reverse, ek_bound
-  real(dp),dimension(1:nvector,1:ndim)::xx
-  real(dp),dimension(1:nvector,1:nvar)::uu
-  real(dp),dimension(1:nvector,1:nvar)::qq
-  real(dp)::dx,rr,vx,vy,vz,pp,eint,ekin,emag,erad
+  real(kind=8)::reverse, ek_bound
+  real(kind=8),dimension(1:nvector,1:ndim)::xx
+  real(kind=8),dimension(1:nvector,1:nvar)::uu
+  real(kind=8),dimension(1:nvector,1:nvar)::qq
+  real(kind=8)::dx,rr,vx,vy,vz,pp,eint,ekin,emag,erad
 
   type = r%bound_type(ibound)
   dir = r%bound_dir(ibound)
