+++ clump/clfind_commons.f90
@@ -4,13 +4,13 @@ module clfind_commons
 
     type clump_t
 
-       real(dp)::relevance_threshold=2
-       real(dp)::density_threshold=-1
-       real(dp)::saddle_threshold=-1
-       real(dp)::mass_threshold=0
-       real(dp)::purity_threshold=-1
-       real(dp)::bound_threshold=2
-       real(dp)::fraction_threshold=0.1d0
+       real(kind=8)::relevance_threshold=2
+       real(kind=8)::density_threshold=-1
+       real(kind=8)::saddle_threshold=-1
+       real(kind=8)::mass_threshold=0
+       real(kind=8)::purity_threshold=-1
+       real(kind=8)::bound_threshold=2
+       real(kind=8)::fraction_threshold=0.1d0
 
        integer :: ntest=0 ! Actual number of test particles in current processor
        integer(kind=8) :: ntest_tot=0 ! Total number of test particles across all processors
@@ -33,24 +33,24 @@ module clfind_commons
        integer,allocatable,dimension(:) :: lev_peak ! AMR level of the peak
        integer,allocatable,dimension(:) :: new_peak ! Peak ID in which current peak has merged
        integer,allocatable,dimension(:) :: npart ! Number of particles per clump
-       real(dp),allocatable,dimension(:) :: max_dens ! Density value at the peak
-       real(dp),allocatable,dimension(:) :: clump_mass ! Mass inside peak patch
-       real(dp),allocatable,dimension(:) :: relevance ! Relevance (prominence) of the peak
-       real(dp),allocatable,dimension(:) :: tidal_dens ! Density at the tidal radius
-       real(dp),allocatable,dimension(:) :: min_dens ! Min density of the clump
-       real(dp),allocatable,dimension(:) :: clump_vol ! Volume of the clump
-       real(dp),allocatable,dimension(:) :: particle_mass ! Clump mass using directly dark matter particles
-       real(dp),allocatable,dimension(:,:) :: peak_pos ! Position of the peak
-       real(dp),allocatable,dimension(:,:) :: peak_vel ! Velocity of the peak
-       real(dp),allocatable,dimension(:,:) :: peak_acc ! Acceleration of the peak
-       real(dp),allocatable,dimension(:,:) :: peak_com ! Center of mass of the peak
+       real(kind=8),allocatable,dimension(:) :: max_dens ! Density value at the peak
+       real(kind=8),allocatable,dimension(:) :: clump_mass ! Mass inside peak patch
+       real(kind=8),allocatable,dimension(:) :: relevance ! Relevance (prominence) of the peak
+       real(kind=8),allocatable,dimension(:) :: tidal_dens ! Density at the tidal radius
+       real(kind=8),allocatable,dimension(:) :: min_dens ! Min density of the clump
+       real(kind=8),allocatable,dimension(:) :: clump_vol ! Volume of the clump
+       real(kind=8),allocatable,dimension(:) :: particle_mass ! Clump mass using directly dark matter particles
+       real(kind=8),allocatable,dimension(:,:) :: peak_pos ! Position of the peak
+       real(kind=8),allocatable,dimension(:,:) :: peak_vel ! Velocity of the peak
+       real(kind=8),allocatable,dimension(:,:) :: peak_acc ! Acceleration of the peak
+       real(kind=8),allocatable,dimension(:,:) :: peak_com ! Center of mass of the peak
 
        integer,allocatable,dimension(:) :: ind_halo ! Peak index of halo patch
        integer,allocatable,dimension(:) :: ind_final ! Final clump index
-       real(dp),allocatable,dimension(:) :: halo_mass ! Mass inside halo patch
-       real(dp),allocatable,dimension(:) :: saddle_dens ! Density of the densest saddle point
-       real(dp),allocatable,dimension(:,:) :: mass_bin ! Cumulative mass profile of halo
-       real(dp),allocatable,dimension(:,:) :: phi ! Gravitational self-potential
+       real(kind=8),allocatable,dimension(:) :: halo_mass ! Mass inside halo patch
+       real(kind=8),allocatable,dimension(:) :: saddle_dens ! Density of the densest saddle point
+       real(kind=8),allocatable,dimension(:,:) :: mass_bin ! Cumulative mass profile of halo
+       real(kind=8),allocatable,dimension(:,:) :: phi ! Gravitational self-potential
 
        integer,allocatable,dimension(:) :: ntree ! Number of merger tree paeticles per clump
        integer,allocatable,dimension(:) :: form_tree ! Does peak form a new tree particle?
