+++ hydro/hydro_parameters.f90
@@ -47,16 +47,16 @@ module const
   use amr_parameters, only: dp
 
   ! Some useful constant
-  real(dp),parameter ::bigreal = 1.0d+30
-  real(dp),parameter ::zero = 0.0
-  real(dp),parameter ::one = 1.0
-  real(dp),parameter ::two = 2.0
-  real(dp),parameter ::three = 3.0
-  real(dp),parameter ::four = 4.0
-  real(dp),parameter ::two3rd = 0.6666666666666667
-  real(dp),parameter ::half = 0.5
-  real(dp),parameter ::third = 0.33333333333333333
-  real(dp),parameter ::forth = 0.25
-  real(dp),parameter ::sixth = 0.16666666666666667
+  real(kind=8),parameter ::bigreal = 1.0d+30
+  real(kind=8),parameter ::zero = 0.0
+  real(kind=8),parameter ::one = 1.0
+  real(kind=8),parameter ::two = 2.0
+  real(kind=8),parameter ::three = 3.0
+  real(kind=8),parameter ::four = 4.0
+  real(kind=8),parameter ::two3rd = 0.6666666666666667
+  real(kind=8),parameter ::half = 0.5
+  real(kind=8),parameter ::third = 0.33333333333333333
+  real(kind=8),parameter ::forth = 0.25
+  real(kind=8),parameter ::sixth = 0.16666666666666667
 
 end module const
